% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_mapping.R
\name{domain_mapping}
\alias{domain_mapping}
\title{domain_mapping}
\usage{
domain_mapping(dataset_desc, dataset, domain)
}
\arguments{
\item{dataset_desc}{The metadata file for the dataset, containing general information. See the demo data file in the data folder of this package.}

\item{dataset}{The metadata file for the dataset, containing information about each variable. See the demo data file in the data folder of this package.}

\item{domain}{The file that lists the domains (latent concepts) of interest to be used within the research study. See the demo data file in the data folder of this package.}
}
\value{
The function will return a log file with your chosen categorizations.
}
\description{
This function will read in the meta-data of a dataset (DataClass) from the SAIL databank, obtained from https://modelcatalogue.cs.ox.ac.uk/hdruk_live/.
It will loop through all the variable names, and ask you to categorise each variable into one of your chosen domains.
The domains (or 'latent concepts') will appear up in the Plots tab for your reference, as well as information about the dataset.
A log file will be saved with the categorizations you made.
}
\examples{
In the R console run these four lines:

data(dataset_desc)
data(dataset)
data(domains)

domain_mapping(dataset_desc,dataset,domains)

It will ask you for a range of rows to process.
For this demo, write '1,10' so that the dataset file can be processed quickly.
If you press enter, it will process all the rows in the dataset file (93).

It will ask you for your initials. This is to save within the log file, to keep track of who made the categorizations.

It will then loop through each row you have requested. It will show you the name, description and data type for each variable.
See the Plots tab for the Domain table, and tables that summarise information about the dataset that the variables derive from.
Respond to this prompt with a single number or multiple numbers separate by a column.
For example 'ALF_E' would be coded as '2' (ID INFORMATION)
The user has an option to write a note explaining their category choice.

These example files (in data) started as excel or csv files.
Use the packages like read.csv or readxl to read your own files into R as a dataframes before calling the domain_mapping function.
}
